requirement.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
Django==3.0.7
djangorestframework==3.11.0
djangotoolbox==1.8.0
future==0.16.0
incremental==17.5.0
ipython-genutils==0.2.0
jsondiff==1.1.1
nose==1.3.7
psycopg2==2.8.5
py==1.8.1
pycparser==2.20
pytest==3.2.1
pytz==2020.1
SQLAlchemy==1.1.12
sqlparse==0.3.1
stevedore==1.32.0
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
